apps/stable-diffusion/src/requirements.txt
diffusers[torch]>=0.24.0
transformers>=4.36.0
accelerate>=0.25.0
pillow>=10.0.0
torch>=2.0.0
safetensors>=0.4.0
gradio>=4.0.0
numpy>=1.24.0